pyproject.toml
[project]
name = "ascii-agent-hackathon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "UFO-101", email = "josephmiller101@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=6.30.1",
    "numpy>=2.3.3",
    "pillow>=11.3.0",
    "requests>=2.32.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
